Workaround fractional scaling completely breaking top_margin positioning
This moves the anchor point to the bottom screen edge if the top_margin
value is greater than 0.5. This makes the issue less noticeable when the
top_margin value is close to 0 or close to 1 (like the default 0.85),
but doesn't fix the issue. The issue is still very apparent when the
top_margin value is 0.5 or close to it.