Port to GTK4 (#105)
* zbus: move to non-deprecated attribute macros

* server: move from MainContext::channel to async_channel

* src/server: fix receiver being closed after ignored libinput event

* server: update gtk to 0.18.1

* server: update to gtk-layer-shell 0.8.1

* initial port to gtk4 (compiles but doesn't work properly)

- some icons are broken
- clearing osd window doesn't work
- gtk4_layer_shell crashes more often than not

* rustfmt.toml: fix warning

* server/osd_window: fix clear_osd() not actually removing things

we need to get the next sibling before we remove it from the box.

* server/osd_window: fix missing icons (has_icon no longer works properly)

some icons (notably volume icons) don't seem to be detected by has_icon
properly in all cases, but the icon is drawn correctly. Use
gio::ThemedIcon directly and pass "missing-symbolic" as a fallback name
through there to avoid checking explicitly.

* server/osd_window: fix gtk4-layer-shell crash

* server/osd_window: update window margin when parameters change

This ensures the window margin is always correct. This also fixes an
annoying issue where the monitor scale factor is not always correct (it
never changes back to scale 1 if it was scale 2 at some point) by
transforming the monitor height into window coordinates manually.

Fixes #101.

* server/application: asynchronously send libinput events

* server: remove unused imports

* server/application: clean up gtk imports